JNR FFI operand error fixed + Verification

+++ src/main/java/jmeos/functions/builder/FunctionsBuilder.java
@@ -38,6 +38,8 @@ public class FunctionsBuilder {
 		HashMap<String, String> typeChange = new HashMap<>();
 		typeChange.put("\\*", "Pointer");
 		typeChange.put("\\*char", "String");
+		//Ce changement fait crash le programme ??
+		//typeChange.put("Pointer[]", "Pointer");
 		typeChange.put("bool", "boolean");
 		typeChange.put("float", "float");
 		typeChange.put("double", "double");
@@ -45,19 +47,19 @@ public class FunctionsBuilder {
 		typeChange.put("int", "int");
 		typeChange.put("short", "short");
 		typeChange.put("long", "long");
-		typeChange.put("int8", "byte");
+		typeChange.put("int8", "short");
 		typeChange.put("int16", "short");
 		typeChange.put("int32", "int");
 		typeChange.put("int64", "long");
-		typeChange.put("int8_t", "byte");
+		typeChange.put("int8_t", "short");
 		typeChange.put("int16_t", "short");
 		typeChange.put("int32_t", "int");
 		typeChange.put("int64_t", "long");
-		typeChange.put("uint8", "byte");
+		typeChange.put("uint8", "short");
 		typeChange.put("uint16", "short");
 		typeChange.put("uint32", "int");
 		typeChange.put("uint64", "long");
-		typeChange.put("uint8_t", "byte");
+		typeChange.put("uint8_t", "short");
 		typeChange.put("uint16_t", "short");
 		typeChange.put("uint32_t", "int");
 		typeChange.put("uint64_t", "long");
